steel-yard-optimizer/requirements.txt
numpy>=1.26
pandas>=2.0
pyyaml>=6.0
jsonschema>=4.20
pyarrow>=14.0
xgboost>=3.1.0
scikit-learn>=1.4
joblib>=1.3
ortools>=9.8
